requirements.txt
qrcode
Pillow
PyQt5